seperating dataset code in dedicated script
